tests(core): fix universal methods for gherkin and new tests
