+++ cypress.json
@@ -3,6 +3,7 @@
   "viewportWidth": 1400,
   "viewportHeight": 800,
   "testFiles": "**/*.feature",
+  "video": false,
   "env": {
     "apiServer": "http://localhost:8000/api/v1/",
     "adminEmail": "test@ergonode.com",
