+++ cypress/integration/common/auth.test.js
