+++ modules/@ergo/attributes/pages/attribute-groups/index.vue
@@ -9,6 +9,7 @@
             :is-read-only="$isReadOnly('ATTRIBUTE')">
             <template #mainAction>
                 <Button
+                    data-cy="newAttributeGroup"
                     title="NEW ATTRIBUTE GROUP"
                     :size="smallSize"
                     :disabled="!$hasAccess(['ATTRIBUTE_CREATE'])"
