+++ modules/@ergo/categories/pages/categories/index.vue
@@ -9,6 +9,7 @@
             :is-read-only="$isReadOnly('CATEGORY')">
             <template #mainAction>
                 <Button
+                    data-cy="newCategory"
                     title="NEW CATEGORY"
                     :size="smallSize"
                     :disabled="!$hasAccess(['CATEGORY_CREATE'])"
