+++ modules/@ergo/category-trees/components/Forms/CategoryTreeForm.vue
@@ -9,6 +9,7 @@
         <template #body="{ errorMessages }">
             <FormSection>
                 <TextField
+                    data-cy="categoryTreeName"
                     :value="code"
                     solid
                     regular
