+++ modules/@ergo/category-trees/pages/category-trees/index.vue
@@ -9,6 +9,7 @@
             :is-read-only="$isReadOnly('CATEGORY_TREE')">
             <template #mainAction>
                 <Button
+                    data-cy="newCategoryTree"
                     title="NEW TREE"
                     :size="smallSize"
                     :disabled="!$hasAccess(['CATEGORY_TREE_CREATE'])"
